+++ app/Http/Controllers/SurveyProgramController.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Controller;
 use App\Http\QueryFilters\MemberFilters;
 use App\Http\QueryFilters\SurveyProgramFilters;
 use App\Http\QueryFilters\UserFilters;
+use App\Http\Requests\MemberChangeRequest;
 use App\Http\Requests\SurveyProgramRequest;
 use App\Http\Resources\SurveyProgramResource;
 use App\Http\Resources\SurveyProgramUserResource;
@@ -21,6 +22,28 @@ use Illuminate\Support\Facades\Auth;
 
 class SurveyProgramController extends Controller
 {
+
+    private function hasPermission($surveyProgram, $permission)
+    {
+        $user_id = Auth::id();
+        $surveyProgramUser = SurveyProgramUser::where('user_id', $user_id)
+            ->where('survey_program_id', $surveyProgram->id)
+            ->where('active', true)
+            ->first();
+
+        if (is_null($surveyProgramUser)) {
+            return false;
+        }
+
+        return $surveyProgramUser->permissions()->where('name', $permission)->count() > 0;
+    }
+
+    public function self(SurveyProgramFilters $filters)
+    {
+        $user = Auth::user();
+        return SurveyProgramResource::collection($user->surveyPrograms()->filterBy($filters)->wherePivot('active', true)->paginate(10));
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -46,7 +69,8 @@ class SurveyProgramController extends Controller
         $mareSurveyProgramUser = SurveyProgramUser::create([
             'survey_program_id' => $surveyProgram->id,
             'user_id' => Auth::id(),
-            'active' => 1
+            'active' => 1,
+            'accepted' => 1,
         ]);
 
         $mareSurveyProgramUser->permissions()->attach(Permission::all()->pluck('id')->toArray());
@@ -75,10 +99,70 @@ class SurveyProgramController extends Controller
      */
     public function update(SurveyProgramRequest $request, SurveyProgram $surveyProgram)
     {
+        if (!self::hasPermission($surveyProgram, 'edit')) {
+            return response()->json(["You don't have access to edit the survey program: \"{$surveyProgram->name}\""], 403);
+        }
+
         $validator = $request->validated();
 
         $surveyProgram->update($validator);
         return new SurveyProgramResource($surveyProgram);
+    }
+
+    public function updateUsers(MemberChangeRequest $request, SurveyProgram $surveyProgram)
+    {
+        if (!self::hasPermission($surveyProgram, 'edit')) {
+            return response()->json(["You don't have access to edit the survey program: \"{$surveyProgram->name}\""], 403);
+        }
+
+        $validator = $request->validated();
+
+        $currentUsersIds = $surveyProgram->surveyProgramUsers->pluck("id")->toArray();
+        $usersToKeep = array();
+
+
+        foreach ($validator["users"] as $user) {
+
+            $existingUser = User::where('email', $user["email"])->first();
+            if (is_null($existingUser)) {
+                return response()->json(
+                    ["message" => "The user with email " . $user["email"] . " doesn't exist"],
+                    409
+                );
+            }
+
+            $surveyProgramUser = SurveyProgramUser::updateOrCreate(
+                [
+                    'user_id' => $existingUser["id"],
+                    "survey_program_id" => $surveyProgram->id
+                ],
+                [
+                    "survey_program_id" => $surveyProgram->id,
+                    "user_id" => $existingUser->id
+                ]
+            );
+            array_push($usersToKeep, $surveyProgramUser->id);
+
+            $surveyProgramUser->permissions()->detach();
+
+            $permissionsToAdd = [Permission::where('name', 'show')->first()->id];
+            if (array_key_exists("create", $user) && $user["create"]) {
+                array_push($permissionsToAdd, Permission::where('name', 'create')->first()->id);
+            }
+
+            if (array_key_exists("edit", $user) && $user["edit"]) {
+                array_push($permissionsToAdd, Permission::where('name', 'edit')->first()->id);
+                array_push($permissionsToAdd, Permission::where('name', 'delete')->first()->id);
+            }
+
+            $surveyProgramUser->permissions()->attach($permissionsToAdd);
+        }
+
+        $idsToRemove = array_diff($currentUsersIds, $usersToKeep);
+
+        SurveyProgramUser::whereIn('id', $idsToRemove)->delete();
+
+        return new SurveyProgramResource($surveyProgram->fresh());
     }
 
     /**
@@ -89,6 +173,11 @@ class SurveyProgramController extends Controller
      */
     public function destroy(SurveyProgram $surveyProgram)
     {
+        if (!self::hasPermission($surveyProgram, 'edit')) {
+            return response()->json(["You don't have access to edit the survey program: \"{$surveyProgram->name}\""], 403);
+        }
+
+        $surveyProgram->surveyProgramUsers()->delete();
         $surveyProgram->delete();
 
         return response()->json(null, 204);
